Create Advanced SSH Script - Lab 6.py
This Lab builds off of the the previous SSH configurations. All of the variables are
reused but the executable script changes. All of the routers will need to be opened
and the “Enter” key should be used on each window a couple times before
shuttingythe windows to execute the script.